refactor: atoms import 경로 수정
